tests/pyproject.toml
[tool.poetry]
name = "kafka-test"
version = "0.1.0"
description = ""
authors = ["Dohyeon Kim <k.dohvis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
